Move groups to shared leaf-level package
