+++ src/go/cmd/local.go
@@ -3,12 +3,14 @@ package cmd
 import (
 	"fmt"
 
+	cmdshared "github.com/kkrull/marmot/cmd/shared"
+
 	"github.com/spf13/cobra"
 )
 
 var localCmd = &cobra.Command{
 	Args:    cobra.NoArgs,
-	GroupID: repositoryGroup.id(),
+	GroupID: cmdshared.RepositoryGroup.Id(),
 	Long:    "Deal with repositories on the local filesystem.",
 	Run: func(cmd *cobra.Command, args []string) {
 		fmt.Println("local called")
